포인트 관련 추가

+++ src/main/java/DB_member/MemberDAO.java
@@ -5,6 +5,9 @@ import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 
+import DB_board.BoardDAO;
+import DB_point.PointDAO;
+
 public class MemberDAO {
 	public static MemberDAO instance = new MemberDAO();
 	
@@ -33,6 +36,38 @@ public class MemberDAO {
 		if (rs != null) {
 			conn.close();
 		}
+	}
+	
+	public int adminCheck(String id, String nick) throws Exception {
+		int check = 0;
+		
+		try {
+			getConn();
+			String sql = "SELECT * FROM admin WHERE admin_id = ?";
+			pstmt = conn.prepareStatement(sql);
+			pstmt.setString(1, id);
+			rs = pstmt.executeQuery();
+			if (rs.next()) {
+				System.out.println("아이디 중복");
+				check = 1; //id같음
+			} else {
+				sql = "SELECT * FROM admin WHERE nickname = ?";
+				pstmt = conn.prepareStatement(sql);
+				pstmt.setString(1, nick);
+				rs = pstmt.executeQuery();
+				if (rs.next()) {
+					System.out.println("닉네임 중복");
+					check = 2; //닉넴같음
+				}	
+			}
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			closeDB();
+		}
+		
+		return check;
 	}
 	
 	public boolean findUser(String id) throws Exception {
@@ -217,30 +252,62 @@ public class MemberDAO {
 		} finally {
 			closeDB();
 		}
+		
+		if (check == 4) {
+			//포인트
+			boolean pointCheck = PointDAO.instance.joinPoint(id);
+			if (pointCheck) {
+				check = 5;
+			}
+		}
+		System.out.println("회원가입 상태 :  " + check);
 		return check;
 	}
 	
 	public boolean delMember(String id) throws Exception {
 		boolean check = false;
 		
-		try {
-			getConn();
-			String sql = "DELETE FROM member WHERE id = ?";
-			pstmt = conn.prepareStatement(sql);
-			pstmt.setString(1, id);
+		boolean boardCheck = false;
+		boolean pointCheck = false;
+		boolean next = false;
+		//게시글 삭제
+		//1. 먼저 작성한거 있는지 확인
+		boardCheck = BoardDAO.instance.postCheck(id);
+		if (boardCheck) {
+			//2. 있으면 삭제 (없으면 굳이)
+			boardCheck = BoardDAO.instance.delMyBoard(id);
 			
-			int result = pstmt.executeUpdate();
-			if (result == 1) {
-				System.out.println("회원 탈퇴 성공");
-				check = true;
-			} else {
-				System.out.println("회원 탈퇴 실패");
+		} else {
+			boardCheck = true;
+		}
+		
+		//포인트 삭제(기록까지)
+		pointCheck = PointDAO.instance.delPointInfo(id);
+		
+		if (boardCheck && pointCheck) {
+			next = true;
+		}
+		
+		if (next) {
+			try {
+				getConn();
+				String sql = "DELETE FROM member WHERE id = ?";
+				pstmt = conn.prepareStatement(sql);
+				pstmt.setString(1, id);
+				
+				int result = pstmt.executeUpdate();
+				if (result == 1) {
+					System.out.println("회원 탈퇴 성공");
+					check = true;
+				} else {
+					System.out.println("회원 탈퇴 실패");
+				}
+				
+			} catch (Exception e) {
+				e.printStackTrace();
+			} finally {
+				closeDB();
 			}
-			
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			closeDB();
 		}
 		
 		return check;
@@ -274,5 +341,25 @@ public class MemberDAO {
 		return check;
 	}
 	
-	
+	public boolean nickCheck(String nick) throws Exception {
+		boolean check = false;
+		
+		try {
+			getConn();
+			String sql = "SELECT * FROM admin WHERE nickname = ?";
+			pstmt = conn.prepareStatement(sql);
+			pstmt.setString(1, nick);
+			rs = pstmt.executeQuery();
+			
+			if (rs.next()) {
+				check = true;
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			closeDB();
+		}
+		
+		return check;
+	}
 }
